bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["bakhteev-v@yandex.ru"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "0.6.0"
alembic = "^1.7.7"
SQLAlchemy = "1.4.28"
python-telegram-bot = "13.9"
psycopg2-binary = "^2.9.3"
minio = "7.1.6"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
ipython = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"